hajat-addons/hajat/requirements.txt
python-decouple==3.7
python-box==6.0.2
